CMPT 276 group project
Coded mostly by Ian Ho and Ethan Jung, with a couple of test cases
written by Henry Fok and a couple lines of code by Brian On, this
back-end storage service utilized Microsoft Azure for storage. It allows
for a list of people alongside their status, friends list, and status
updates to be listed as JSON objects, with the partition and row being
the country and name respectively. A number of operations can be
performed, including PUSH, GET and POST. An Authentification server
checks and gives out a token to check whether or not a user is logged
in, and the a UserServer and PushServer performs requests to both the
Authentification and Basic Server to perform requests like adding
friends or changing one's status.